added kiss_fft_next_fast_size() to determine the next number divisible by the radices 2,3,5

+++ tools/kiss_fftnd.c
@@ -73,7 +73,7 @@ kiss_fftnd_cfg kiss_fftnd_alloc(const int *dims,int ndims,int inverse_fft,void*m
         st->states[i] = kiss_fft_alloc (st->dims[i], inverse_fft, ptr,&len);
         ptr += len;
     }
-    if ( ptr - (char*)st != memneeded ) {
+    if ( ptr - (char*)st != (int)memneeded ) {
         fprintf(stderr,
                 "################################################################################\n"
                 "Internal error! Memory allocation miscalculation\n"
